# docs/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FitTrack API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/scale-sync/v2": {
      "post": {
        "tags": [
          "scale-sync"
        ],
        "summary": "Ingest Scale Measurement",
        "operationId": "ingest_scale_measurement_api_scale_sync_v2_post",
        "parameters": [
          {
            "name": "x-fittrack-device-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Fittrack-Device-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScaleSyncV2Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/scale-measurements": {
      "get": {
        "tags": [
          "scale-measurements"
        ],
        "summary": "List Measurements",
        "operationId": "list_measurements_api_scale_measurements_get",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/scale-measurements/{measurement_id}/reject": {
      "post": {
        "tags": [
          "scale-measurements"
        ],
        "summary": "Reject Measurement",
        "operationId": "reject_measurement_api_scale_measurements__measurement_id__reject_post",
        "parameters": [
          {
            "name": "measurement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Measurement Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/account/body-profile": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "Get Body Profile",
        "operationId": "get_body_profile_api_account_body_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/BodyProfileResponse"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Body Profile Api Account Body Profile Get"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "account"
        ],
        "summary": "Update Body Profile",
        "operationId": "update_body_profile_api_account_body_profile_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BodyProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BodyProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BodyProfileResponse": {
        "properties": {
          "height_cm": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height Cm"
          },
          "birth_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Date"
          },
          "calculation_sex": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "male",
                  "female"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Calculation Sex"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "BodyProfileResponse"
      },
      "BodyProfileUpdate": {
        "properties": {
          "height_cm": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 300.0,
                "minimum": 50.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height Cm"
          },
          "birth_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Date"
          },
          "calculation_sex": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "male",
                  "female"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Calculation Sex"
          }
        },
        "type": "object",
        "title": "BodyProfileUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ScaleSyncV2Request": {
        "properties": {
          "device_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Device Id"
          },
          "device_event_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Device Event Id"
          },
          "measured_at": {
            "type": "string",
            "format": "date-time",
            "title": "Measured At"
          },
          "weight_kg": {
            "type": "number",
            "maximum": 300.0,
            "minimum": 0.5,
            "title": "Weight Kg"
          },
          "impedance_ohm": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 5000.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Impedance Ohm"
          },
          "protocol": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Protocol"
          },
          "protocol_version": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Protocol Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "device_id",
          "device_event_id",
          "measured_at",
          "weight_kg",
          "protocol",
          "protocol_version"
        ],
        "title": "ScaleSyncV2Request",
        "description": "Raw device-only payload. Extra fields are rejected deliberately."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}